+++ src/components/shared/formItems/index.tsx
@@ -12,7 +12,7 @@ import ProjectPayButton from './ProjectPayButton'
 import ProjectPayDisclosure from './ProjectPayDisclosure'
 import ProjectReserved from './ProjectReserved'
 import ProjectTarget from './ProjectTarget'
-import ProjectTicketMods from '../../v1/ProjectTicketMods'
+import ProjectTicketMods from '../../v1/shared/ProjectTicketMods'
 import ProjectTwitter from './ProjectTwitter'
 import TokenRefs from '../../v1/shared/formItems/TokenRefs'
 import { ItemNoInput } from './ItemNoInput'
