+++ src/components/v1/shared/FundingCycle/modals/DistributeTokensModal.tsx
@@ -1,7 +1,7 @@
 import { Trans } from '@lingui/macro'
 import { Modal, Space } from 'antd'
 import FormattedAddress from 'components/shared/FormattedAddress'
-import TicketModsList from 'components/v1/TicketModsList'
+import TicketModsList from 'components/v1/shared/TicketModsList'
 import { V1ProjectContext } from 'contexts/v1/projectContext'
 import useReservedTokensOfProject from 'hooks/v1/contractReader/ReservedTokensOfProject'
 import { useDistributeTokensTx } from 'hooks/v1/transactor/DistributeTokensTx'
