+++ src/components/v1/shared/formItems/ProjectHandle/ProjectHandleFormItem.tsx
@@ -8,12 +8,12 @@ import { ThemeContext } from 'contexts/themeContext'
 import useContractReader from 'hooks/v1/contractReader/ContractReader'
 import { V1ContractName } from 'models/v1/contracts'
 import { normalizeHandle } from 'utils/formatHandle'
+import { FormItemExt } from 'components/shared/formItems/formItemExt'
 
 import {
   ProjectHandleInput,
   ProjectHandleInitialValue,
 } from './ProjectHandleInput'
-import { FormItemExt } from '../../../../shared/formItems/formItemExt'
 
 /**
  * Custom Form.Item component for project handles.
