+++ src/components/v1/shared/forms/PayModsForm.tsx
@@ -11,7 +11,7 @@ import { fromWad, perbicentToPercent } from 'utils/formatNumber'
 
 import { getTotalPercentage } from 'components/shared/formItems/formHelpers'
 import { CurrencyContext } from 'contexts/currencyContext'
-import ProjectPayoutMods from 'components/v1/ProjectPayMods/ProjectPayoutMods'
+import ProjectPayoutMods from 'components/v1/shared/ProjectPayMods/ProjectPayoutMods'
 
 import * as constants from '@ethersproject/constants'
